+++ db/migrate/20260624120000_add_bank_card_id_to_mobile_payment_methods.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# B1.12-R1 rev2: CardId из ответа Т-Банка (UserCards.card_id в ТЗ).
+class AddBankCardIdToMobilePaymentMethods < ActiveRecord::Migration[8.0]
+  def change
+    add_column :mobile_payment_methods, :bank_card_id, :string, limit: 64, if_not_exists: true
+  end
+end